Spatial points representing home locations

These points represent population-weighted centroids of Medium Super Output Area (MSOA) zones within a 1 mile radius of of my home when I was writing this package.

data(cents)

Format

A spatial dataset with 8 rows and 5 variables

Details

geo_code the official code of the zone

MSOA11NM name zone name

percent_fem the percent female

avslope average gradient of the zone

Cents was generated from the data repository pct-data: https://github.com/npct/pct-data. This data was accessed from within the pct repo: https://github.com/npct/pct, using the following code:

Examples

# NOT RUN {
cents
plot(cents)
# }