Clip the first and last n metres of SpatialLines

Takes lines and removes the start and end point, to a distance determined by the user.

geo_toptail(l, toptail_dist, ...)

Arguments

l
A SpatialLines object

toptail_dist
The distance (in metres) to top and tail the line by. Can either be a single value or a vector of the same length as the SpatialLines object.

...
Arguments passed to rgeos::gBuffer()

Details

Note: toptailgs() is around 10 times faster, but only works on data with geographic CRS's due to its reliance on the geosphere package.

Examples

l <- routes_fast[2:4, ]
l_toptail <- geo_toptail(l, toptail_dist = 300)

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

plot(l)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(l_toptail, col = "red", add = TRUE, lwd = 3)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(cents, col = "blue", add = TRUE, pch = 15)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

# Note the behaviour when the buffer size removes lines
r_toptail <- geo_toptail(l, toptail_dist = 900)

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Line 1 is completely removed by the clip and is omitted from the results

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Line 2 is completely removed by the clip and is omitted from the results

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

plot(r_toptail, lwd = 9, add = TRUE) # short route removed

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

geo_toptail(routes_fast_sf[2:4, ], 300)

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Simple feature collection with 3 features and 0 fields
#> geometry type:  MULTILINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.546641 ymin: 53.81764 xmax: -1.512596 ymax: 53.83041
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>                         geometry
#> 1 MULTILINESTRING ((-1.521191...
#> 2 MULTILINESTRING ((-1.521191...
#> 3 MULTILINESTRING ((-1.516117...