Function to split overlapping SpatialLines into segments

Divides SpatialLinesDataFrame objects into separate Lines. Each new Lines object is the aggregate of a single number of aggregated lines.

gsection(sl, buff_dist = 0)

Arguments

sl
SpatialLinesDataFrame with overlapping Lines to split by number of overlapping features.

buff_dist
A number specifying the distance in meters of the buffer to be used to crop lines before running the operation. If the distance is zero (the default) touching but non-overlapping lines may be aggregated.

Examples

sl <- routes_fast[2:4, ]
rsec <- gsection(sl)
rsec_buff <- gsection(sl, buff_dist = 1)

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

#> Transforming to CRS +proj=aeqd +lat_0=53.825992 +lon_0=-1.52795833 +x_0=0 +y_0=0 +ellps=WGS84

plot(sl[1], lwd = 9, col = 1:nrow(sl))

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(rsec, col = 5 + (1:length(rsec)), add = TRUE, lwd = 3)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(rsec_buff, col = 5 + (1:length(rsec_buff)), add = TRUE, lwd = 3)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'
# NOT RUN {
sl <- routes_fast_sf[2:4, ]
rsec <- gsection(sl)
rsec <- gsection(sl, buff_dist = 100) # 4 features: issue
# }