Return SpatialPointsDataFrame with located points from OSRM locate service

Return SpatialPointsDataFrame with located points from OSRM locate service

locate2spdf(lat, lng = lng, osrmurl = "http://router.project-osrm.org",
  return_sf = FALSE)

Arguments

lat
Numeric vector containing latitude coordinate for each coordinate to map. Also accepts dataframe with latitude in the first column and longitude in the second column.

lng
Numeric vector containing longitude coordinate for each coordinate to map.

osrmurl
Base URL of the OSRM service

return_sf
Boolean value if this function should return an sf object, if FALSE returns sp object (default FALSE).

Details

Retrieve coordinates of the node(s) on the network mapped from coordinates passed to functions using OSRM API v4 only. For API v5, use nearest_osm.

See also

Examples

# NOT RUN {
locate2spdf(
  lat = c(50.3, 50.2),
  lng = c(13.2, 13.1)
)
# }