Generate nearest point on the route network of a point using the Google Maps API

Generate nearest point on the route network of a point using the Google Maps API

nearest_google(lat, lng, google_api)

Arguments

lat
Numeric vector containing latitude coordinate for each coordinate to map. Also accepts dataframe with latitude in the first column and longitude in the second column.

lng
Numeric vector containing longitude coordinate for each coordinate to map.

google_api
String value containing the Google API key to use.

Details

Retrieve coordinates of the node(s) on the network mapped from coordinates passed to functions.

Examples

not_run({
   nearest_google(lat = 50.333, lng = 3.222, google_api = "api_key_here")
})