Extract coordinates from OD data

Extract coordinates from OD data

od2odf(flow, zones)

Arguments

flow
A data frame representing the flow between two points or zones. The first two columns of this data frame should correspond to the first column of the data in the zones. Thus in cents, the first column is geo_code. This corresponds to the first two columns of flow.

zones
A SpatialPolygonsDataFrame or SpatialPointsDataFrame representing origins and destinations of travel flows.

Details

Origin-destination ('OD') flow data is often provided in the form of 1 line per flow with zone codes of origin and destination centroids. This can be tricky to plot and link-up with geographical data. This function makes the task easier.

References

Rae, A. (2009). From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census. Computers, Environment and Urban Systems, 33(3). doi:10.1016/j.compenvurbsys.2009.01.007

Examples

data(flow)
data(zones)
od2odf(flow, zones)

#>       code_o    code_d        fx       fy        tx       ty
#> 1  E02002361 E02002361 -1.514962 53.82911 -1.514962 53.82911
#> 2  E02002361 E02002363 -1.514962 53.82911 -1.535735 53.82863
#> 3  E02002361 E02002367 -1.514962 53.82911 -1.552309 53.82317
#> 4  E02002361 E02002371 -1.514962 53.82911 -1.530837 53.81888
#> 5  E02002361 E02002377 -1.514962 53.82911 -1.518509 53.81526
#> 6  E02002361 E02002382 -1.514962 53.82911 -1.512147 53.81101
#> 7  E02002361 E02002393 -1.514962 53.82911 -1.523412 53.80440
#> 8  E02002363 E02002361 -1.535735 53.82863 -1.514962 53.82911
#> 9  E02002363 E02002363 -1.535735 53.82863 -1.535735 53.82863
#> 10 E02002363 E02002367 -1.535735 53.82863 -1.552309 53.82317
#> 11 E02002363 E02002371 -1.535735 53.82863 -1.530837 53.81888
#> 12 E02002363 E02002377 -1.535735 53.82863 -1.518509 53.81526
#> 13 E02002363 E02002382 -1.535735 53.82863 -1.512147 53.81101
#> 14 E02002363 E02002393 -1.535735 53.82863 -1.523412 53.80440
#> 15 E02002367 E02002361 -1.552309 53.82317 -1.514962 53.82911
#> 16 E02002367 E02002363 -1.552309 53.82317 -1.535735 53.82863
#> 17 E02002367 E02002367 -1.552309 53.82317 -1.552309 53.82317
#> 18 E02002367 E02002371 -1.552309 53.82317 -1.530837 53.81888
#> 19 E02002367 E02002377 -1.552309 53.82317 -1.518509 53.81526
#> 20 E02002367 E02002382 -1.552309 53.82317 -1.512147 53.81101
#> 21 E02002367 E02002393 -1.552309 53.82317 -1.523412 53.80440
#> 22 E02002371 E02002361 -1.530837 53.81888 -1.514962 53.82911
#> 23 E02002371 E02002363 -1.530837 53.81888 -1.535735 53.82863
#> 24 E02002371 E02002367 -1.530837 53.81888 -1.552309 53.82317
#> 25 E02002371 E02002371 -1.530837 53.81888 -1.530837 53.81888
#> 26 E02002371 E02002377 -1.530837 53.81888 -1.518509 53.81526
#> 27 E02002371 E02002382 -1.530837 53.81888 -1.512147 53.81101
#> 28 E02002371 E02002393 -1.530837 53.81888 -1.523412 53.80440
#> 29 E02002377 E02002361 -1.518509 53.81526 -1.514962 53.82911
#> 30 E02002377 E02002363 -1.518509 53.81526 -1.535735 53.82863
#> 31 E02002377 E02002367 -1.518509 53.81526 -1.552309 53.82317
#> 32 E02002377 E02002371 -1.518509 53.81526 -1.530837 53.81888
#> 33 E02002377 E02002377 -1.518509 53.81526 -1.518509 53.81526
#> 34 E02002377 E02002382 -1.518509 53.81526 -1.512147 53.81101
#> 35 E02002377 E02002393 -1.518509 53.81526 -1.523412 53.80440
#> 36 E02002382 E02002361 -1.512147 53.81101 -1.514962 53.82911
#> 37 E02002382 E02002363 -1.512147 53.81101 -1.535735 53.82863
#> 38 E02002382 E02002367 -1.512147 53.81101 -1.552309 53.82317
#> 39 E02002382 E02002371 -1.512147 53.81101 -1.530837 53.81888
#> 40 E02002382 E02002377 -1.512147 53.81101 -1.518509 53.81526
#> 41 E02002382 E02002382 -1.512147 53.81101 -1.512147 53.81101
#> 42 E02002382 E02002393 -1.512147 53.81101 -1.523412 53.80440
#> 43 E02002393 E02002361 -1.523412 53.80440 -1.514962 53.82911
#> 44 E02002393 E02002363 -1.523412 53.80440 -1.535735 53.82863
#> 45 E02002393 E02002367 -1.523412 53.80440 -1.552309 53.82317
#> 46 E02002393 E02002371 -1.523412 53.80440 -1.530837 53.81888
#> 47 E02002393 E02002377 -1.523412 53.80440 -1.518509 53.81526
#> 48 E02002393 E02002382 -1.523412 53.80440 -1.512147 53.81101
#> 49 E02002393 E02002393 -1.523412 53.80440 -1.523412 53.80440