Convert a series of points into a dataframe of origins and destinations

Takes a series of geographical points and converts them into a data.frame representing the potential flows, or 'spatial interaction', between every combination of points.

points2odf(p)

Arguments

p
SpatialPointsDataFrame or data.frame

Examples

data(cents)
df <- points2odf(cents)
cents_centroids <- rgeos::gCentroid(cents, byid = TRUE)
df2 <- points2odf(cents_centroids)