Import and format UK 'Stats19' road traffic casualty data

Import and format UK 'Stats19' road traffic casualty data

read_stats19_ve(data_dir = tempdir(), filename = "Vehicles0514.csv")

Arguments

data_dir
Character string representing where the data is stored. If empty, R will attempt to download and unzip the data for you.

filename
Character string of the filename of the .csv to read in - default values are those downloaded from the UK Department for Transport (DfT).

Details

This is a wrapper function to access and load stats 19 data in a user-friendly way. The function returns a data frame, in which each record is a reported incident in the stats19 dataset.

Ensure you have a fast internet connection and at least 100 Mb space.

Examples

not_run({
  ve <- read_stats19_ve()
})