An S4 class representing a (typically) transport network

This class uses a combination of a sf layer and an igraph object to represent transport networks that can be used for routing and other network analyses.

Slots

sl

A sf line layer with the geometry and other attributes for each link the in network.

g

The graph network corresponding to sl.

nb

A list containing vectors of the nodes connected to each node in the network.

weightfield

A character vector containing the variable (column) name from the SpatialLinesDataFrame to be used for weighting the network.