stplanr: Sustainable Transport Planning with R

The stplanr package provides functions to access and analyse data for transportation research, including origin-destination analysis, route allocation and modelling travel patterns.

Interesting functions

overline() - Aggregate overlaying route lines and data intelligently

calc_catchment() - Create a 'catchment area' to show the areas serving a destination

route_cyclestreet() - Finds the fastest routes for cyclists between two places.

See also