Combines two ID values to create a single ID number

Combines two ID values to create a single ID number

szudzik_pairing(val1, val2, ordermatters = FALSE)

Arguments

val1
a vector of numeric, character, or factor values

val2
a vector of numeric, character, or factor values

ordermatters
logical, does the order of values matter to pairing, default = FALSE

Details

In OD data it is common to have many flows from "A to B" and "B to A". It can be useful to group these an have a single ID that represents pairs of IDs with or without directionality.

This function implements the Szudzik pairing function, on two vectors of equal length. It returns a vector of ID numbers.

This function superseeds od_id_order as it is faster on large datasets

Examples

szudzik_pairing(od_data_sample[[1]], od_data_sample[[2]])

#>  [1]  2  5 10 17 26 37 50 65  5  6 11 18 27 38 51 66 10 11 12 19 28 39 52 67 17
#> [26] 18 19 20 29 40 53 68 26 27 28 29 30 41 54 69 37 38 39 40 41 42 55 70 50 51
#> [51] 52 53 54 55 56 71 65 66 67 68 69 70 71 72

od_id_order(od_data_sample)

#>    stplanr.id1 stplanr.id1.1         stplanr.key
#> 1    E02002361     E02002361 E02002361 E02002361
#> 2    E02002361     E02002363 E02002361 E02002363
#> 3    E02002361     E02002367 E02002361 E02002367
#> 4    E02002361     E02002371 E02002361 E02002371
#> 5    E02002361     E02002377 E02002361 E02002377
#> 6    E02002361     E02002382 E02002361 E02002382
#> 7    E02002361     E02002384 E02002361 E02002384
#> 8    E02002361     E02002393 E02002361 E02002393
#> 9    E02002363     E02002361 E02002361 E02002363
#> 10   E02002363     E02002363 E02002363 E02002363
#> 11   E02002363     E02002367 E02002363 E02002367
#> 12   E02002363     E02002371 E02002363 E02002371
#> 13   E02002363     E02002377 E02002363 E02002377
#> 14   E02002363     E02002382 E02002363 E02002382
#> 15   E02002363     E02002384 E02002363 E02002384
#> 16   E02002363     E02002393 E02002363 E02002393
#> 17   E02002367     E02002361 E02002361 E02002367
#> 18   E02002367     E02002363 E02002363 E02002367
#> 19   E02002367     E02002367 E02002367 E02002367
#> 20   E02002367     E02002371 E02002367 E02002371
#> 21   E02002367     E02002377 E02002367 E02002377
#> 22   E02002367     E02002382 E02002367 E02002382
#> 23   E02002367     E02002384 E02002367 E02002384
#> 24   E02002367     E02002393 E02002367 E02002393
#> 25   E02002371     E02002361 E02002361 E02002371
#> 26   E02002371     E02002363 E02002363 E02002371
#> 27   E02002371     E02002367 E02002367 E02002371
#> 28   E02002371     E02002371 E02002371 E02002371
#> 29   E02002371     E02002377 E02002371 E02002377
#> 30   E02002371     E02002382 E02002371 E02002382
#> 31   E02002371     E02002384 E02002371 E02002384
#> 32   E02002371     E02002393 E02002371 E02002393
#> 33   E02002377     E02002361 E02002361 E02002377
#> 34   E02002377     E02002363 E02002363 E02002377
#> 35   E02002377     E02002367 E02002367 E02002377
#> 36   E02002377     E02002371 E02002371 E02002377
#> 37   E02002377     E02002377 E02002377 E02002377
#> 38   E02002377     E02002382 E02002377 E02002382
#> 39   E02002377     E02002384 E02002377 E02002384
#> 40   E02002377     E02002393 E02002377 E02002393
#> 41   E02002382     E02002361 E02002361 E02002382
#> 42   E02002382     E02002363 E02002363 E02002382
#> 43   E02002382     E02002367 E02002367 E02002382
#> 44   E02002382     E02002371 E02002371 E02002382
#> 45   E02002382     E02002377 E02002377 E02002382
#> 46   E02002382     E02002382 E02002382 E02002382
#> 47   E02002382     E02002384 E02002382 E02002384
#> 48   E02002382     E02002393 E02002382 E02002393
#> 49   E02002384     E02002361 E02002361 E02002384
#> 50   E02002384     E02002363 E02002363 E02002384
#> 51   E02002384     E02002367 E02002367 E02002384
#> 52   E02002384     E02002371 E02002371 E02002384
#> 53   E02002384     E02002377 E02002377 E02002384
#> 54   E02002384     E02002382 E02002382 E02002384
#> 55   E02002384     E02002384 E02002384 E02002384
#> 56   E02002384     E02002393 E02002384 E02002393
#> 57   E02002393     E02002361 E02002361 E02002393
#> 58   E02002393     E02002363 E02002363 E02002393
#> 59   E02002393     E02002367 E02002367 E02002393
#> 60   E02002393     E02002371 E02002371 E02002393
#> 61   E02002393     E02002377 E02002377 E02002393
#> 62   E02002393     E02002382 E02002382 E02002393
#> 63   E02002393     E02002384 E02002384 E02002393
#> 64   E02002393     E02002393 E02002393 E02002393

n = 1000
ids <- as.character(runif(n, 1e4, 1e7 - 1))
x <- data.frame(id1 = rep(ids, times = n),
                id2 = rep(ids, each = n),
                val = 1,
                stringsAsFactors = FALSE)
system.time(od_id_order(x))

#>    user  system elapsed 
#>   1.528   0.020   1.547 

system.time(szudzik_pairing(x$id1, x$id2))

#>    user  system elapsed 
#>   0.192   0.012   0.204