Find graph node ID of closest node to given coordinates

Find graph node ID of closest node to given coordinates

find_network_nodes(sln, x, y = NULL, maxdist = 1000)

Arguments

sln
SpatialLinesNetwork to search.

x
Either the x (longitude) coordinate value, a vector of x values, a dataframe or matrix with (at least) two columns, the first for coordinate for x (longitude) values and a second for y (latitude) values, or a named vector of length two with values of 'lat' and 'lon'. The output of geo_code() either as a single result or as multiple (using rbind() ) can also be used.

y
Either the y (latitude) coordinate value or a vector of y values.

maxdist
The maximum distance within which to match the nodes to coordinates. If the SpatialLinesNetwork is projected then distance should be in the same units as the projection. If longlat, then distance is in metres. Default is 1000.

Value

An integer value with the ID of the node closest to (x,y) with a value of NA the closest node is further than maxdist from (x,y). If x is a vector, returns a vector of Node IDs.

Details

Finds the node ID of the closest point to a single coordinate pair (or a set of coordinates) from a SpatialLinesNetwork.

Examples

data(routes_fast)
rnet <- overline(routes_fast, attrib = "length")
SLN <- SpatialLinesNetwork(rnet)
find_network_nodes(SLN, -1.516734, 53.828)

#> [1] 27