Extract coordinates from OD data

Extract coordinates from OD data

od2odf(flow, zones)

Arguments

flow
A data frame representing the flow between two points or zones. The first two columns of this data frame should correspond to the first column of the data in the zones. Thus in cents(), the first column is geo_code. This corresponds to the first two columns of flow().

zones
A spatial object representing origins (and destinations if no separate destinations object is provided) of travel.

Details

Origin-destination ('OD') flow data is often provided in the form of 1 line per flow with zone codes of origin and destination centroids. This can be tricky to plot and link-up with geographical data. This function makes the task easier.

See also

Examples

data(flow)
data(zones)
od2odf(flow[1:2, ], zones)

#>      code_o    code_d        fx       fy        tx       ty
#> 1 E02002361 E02002361 -1.514962 53.82911 -1.514962 53.82911
#> 2 E02002361 E02002363 -1.514962 53.82911 -1.535735 53.82863