Create matrices representing origin-destination coordinates

This function takes a wide range of input data types (spatial lines, points or text strings) and returns a matrix of coordinates representing origin (fx, fy) and destination (tx, ty) points.

od_coords(from = NULL, to = NULL, l = NULL)

Arguments

from
An object representing origins (if lines are provided as the first argument, from is assigned to l)

to
An object representing destinations

l
Only needed if from and to are empty, in which case this should be a spatial object representing desire lines

See also

Examples

od_coords(from = c(0, 52), to = c(1, 53)) # lon/lat coordinates

#>      fx fy tx ty
#> [1,]  0 52  1 53

od_coords(from = cents[1, ], to = cents[2, ]) # Spatial points

#>             fx       fy        tx       ty
#> [1,] -1.546463 53.80952 -1.511861 53.81161

od_coords(cents_sf[1:3, ], cents_sf[2:4, ]) # sf points

#>          fx       fy        tx       ty
#> 1 -1.546463 53.80952 -1.511861 53.81161
#> 2 -1.511861 53.81161 -1.524205 53.80410
#> 3 -1.524205 53.80410 -1.550806 53.82442

# od_coords("Hereford", "Leeds") # geocode locations
od_coords(flowlines[1:3, ])

#>             fx       fy        tx       ty
#> [1,] -1.516734 53.82887 -1.516734 53.82887
#> [2,] -1.516734 53.82887 -1.535617 53.82847
#> [3,] -1.516734 53.82887 -1.550807 53.82442

od_coords(flowlines_sf[1:3, ])

#>             fx       fy        tx       ty
#> [1,] -1.516734 53.82887 -1.516734 53.82887
#> [2,] -1.516734 53.82887 -1.535617 53.82847
#> [3,] -1.516734 53.82887 -1.550807 53.82442