Combine two ID values to create a single ID number

Combine two ID values to create a single ID number

od_id_szudzik(x, y, ordermatters = FALSE)

od_id_max_min(x, y)

od_id_character(x, y)

Arguments

x
a vector of numeric, character, or factor values

y
a vector of numeric, character, or factor values

ordermatters
logical, does the order of values matter to pairing, default = FALSE

Details

In OD data it is common to have many 'oneway' flows from "A to B" and "B to A". It can be useful to group these an have a single ID that represents pairs of IDs with or without directionality, so they contain 'twoway' or bi-directional values.

od_id* functions take two vectors of equal length and return a vector of IDs, which are unique for each combination but the same for twoway flows.

the Szudzik pairing function, on two vectors of equal length. It returns a vector of ID numbers.

This function superseeds od_id_order as it is faster on large datasets

Examples

d <- od_data_sample[2:9, 1:2]
d

#> # A tibble: 8 x 2
#>   geo_code1 geo_code2
#>   <chr>     <chr>    
#> 1 E02002361 E02002363
#> 2 E02002361 E02002367
#> 3 E02002361 E02002371
#> 4 E02002361 E02002377
#> 5 E02002361 E02002382
#> 6 E02002361 E02002384
#> 7 E02002361 E02002393
#> 8 E02002363 E02002361

id <- od_id_character(d[[1]], d[[2]])
id

#> [1] "E02002361 E02002363" "E02002361 E02002367" "E02002361 E02002371"
#> [4] "E02002361 E02002377" "E02002361 E02002382" "E02002361 E02002384"
#> [7] "E02002361 E02002393" "E02002361 E02002363"

duplicated(id)

#> [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE

od_id_szudzik(d[[1]], d[[2]])

#> [1]  5 10 17 26 37 50 65  5

od_id_max_min(d[[1]], d[[2]])

#> [1]  4  7 11 16 22 29 37  4

n <- 1000
ids <- as.character(runif(n, 1e4, 1e7 - 1))
# benchmark of methods:
x <- data.frame(
  id1 = rep(ids, times = n),
  id2 = rep(ids, each = n),
  val = 1,
  stringsAsFactors = FALSE
)
bench::mark(
  check = FALSE,
  od_id_order(x),
  od_id_character(x$id1, x$id2),
  od_id_szudzik(x$id1, x$id2),
  od_id_max_min(x$id1, x$id2)
)

#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.

#> # A tibble: 4 x 13
#>   expression                      min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                    <bch> <bch:>     <dbl> <bch:byt>    <dbl> <int>
#> 1 od_id_order(x)                998ms  998ms      1.00    29.9MB     1.00     1
#> 2 od_id_character(x$id1, x$id2) 953ms  953ms      1.05    22.9MB     0        1
#> 3 od_id_szudzik(x$id1, x$id2)   248ms  281ms      3.56     146MB     8.91     2
#> 4 od_id_max_min(x$id1, x$id2)   196ms  196ms      5.10    84.8MB    10.2      3
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>