Plan a single route with TransportAPI.com

Provides an R interface to the TransportAPI.com public transport API. The function returns a SpatialLinesDataFrame object representing the public route. Currently only works for the United Kingdom. See https://developer.transportapi.com/documentationfor more information.

route_transportapi_public(from, to, silent = FALSE,
  region = "southeast", modes = NA, not_modes = NA)

Arguments

from
Text string or coordinates (a numeric vector of length = 2 representing latitude and longitude) representing a point on Earth.

to
Text string or coordinates (a numeric vector of length = 2 representing latitude and longitude) representing a point on Earth. This represents the destination of the trip.

silent
Logical (default is FALSE). TRUE hides request sent.

region
String for the active region to use for journey plans. Possible values are 'southeast' (default) or 'tfl'.

modes
Vector of character strings containing modes to use. Default is to use all modes.

not_modes
Vector of character strings containing modes not to use. Not used if modes is set.

Details

This function uses the online routing service TransportAPI.com to find public routes between origins and destinations. It does not require any key to access the API.

Note that if from and to are supplied as character strings (instead of lon/lat pairs), Google's geo-coding services are used via geo_code.

See also

Examples

# NOT RUN {
# Plan the 'public' route from Hereford to Leeds
rqh <- route_transportapi_public(from = "Hereford", to = "Leeds")
plot(rq_hfd)
# }

# Aim plan public transport routes with transportAPI